Trying to add ssl certificate to my domain.

+++ certbot/generate-certificate.sh
@@ -1,21 +1,23 @@
 #!/bin/sh
 
-# This script generates SSL certificate
-
-# Clean folder, where could be old certificates
+# Убедитесь, что предыдущие сертификаты удалены
 rm -rf /etc/letsencrypt/live/certfolder*
 
-# Give ourselves certificate
+# Запустите certbot с опцией --non-interactive
 certbot certonly --standalone \
+    --webroot-path /var/www/certbot \
+    --non-interactive \
     --email $DOMAIN_EMAIL \
     -d $DOMAIN_URL \
     --cert-name=certfolder \
     --key-type rsa \
     --agree-tos
 
-# Removing old certificates from mounted nginx folder.
-rm /etc/nginx/cert.pem /etc/nginx/key.pem
-
-# Copy new certificates from certbot to mounted nginx folder
-cp /etc/letsencrypt/live/certfolder/fullchain.pem /etc/nginx/cert.pem
-cp /etc/letsencrypt/live/certfolder/privkey.pem /etc/nginx/key.pem
+# Проверка успешности получения сертификатов
+if [ -f /etc/letsencrypt/live/certfolder/fullchain.pem ] && [ -f /etc/letsencrypt/live/certfolder/privkey.pem ]; then
+    # Копируем сертификаты в папку Nginx
+    cp /etc/letsencrypt/live/certfolder/fullchain.pem /etc/nginx/cert.pem
+    cp /etc/letsencrypt/live/certfolder/privkey.pem /etc/nginx/key.pem
+else
+    echo "Сертификаты не были сгенерированы. Проверьте настройки."
+fi
